完成RBAC功能

+++ backend/controllers/PostController.php
@@ -5,7 +5,9 @@ namespace backend\controllers;
 use Yii;
 use common\models\Post;
 use common\models\PostSearch;
+use yii\filters\AccessControl;
 use yii\web\Controller;
+use yii\web\ForbiddenHttpException;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
 
@@ -25,6 +27,22 @@ class PostController extends Controller
                 'actions' => [
                     'delete' => ['POST'],
                 ],
+            ],
+
+            'access'=>[
+                'class'=>AccessControl::className(),
+                'rules'=>[
+                    [
+                        'actions'=>['index','view'],
+                        'allow'=>true,
+                        'roles'=>['?'],
+                    ],
+                    [
+                        'actions'=>['view','index','create','update'],
+                        'allow'=>true,
+                        'roles'=>['@'],
+                    ],
+                ]
             ],
         ];
     }
@@ -66,9 +84,13 @@ class PostController extends Controller
      * Creates a new Post model.
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
+     * @throws ForbiddenHttpException
      */
     public function actionCreate()
     {
+        if (!Yii::$app->user->can('createPost')){
+            throw new ForbiddenHttpException("对不起，你没有进行该操作的权限");
+        }
         $model = new Post();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
@@ -88,6 +110,9 @@ class PostController extends Controller
      */
     public function actionUpdate($id)
     {
+        if (!Yii::$app->user->can('updatePost')){
+            throw new ForbiddenHttpException("对不起，你没有进行该操作的权限");
+        }
         $model = $this->findModel($id);
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->id]);
@@ -106,6 +131,9 @@ class PostController extends Controller
      */
     public function actionDelete($id)
     {
+        if (!Yii::$app->user->can('deletePost')){
+            throw new ForbiddenHttpException("对不起，你没有进行该操作的权限");
+        }
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
